backend/requirements.txt
fastapi
uvicorn
motor
pydantic
transformers
torch
sentencepiece
python-dotenv
protobuf
# 4vmPQousjn5JJxlg